+++ src/types/source_routing_header.rs
@@ -1,0 +1,4 @@
+pub type NodeId = u8;
+
+// False types to make cargo run happy
+pub type SourceRoutingHeader = [NodeId; 16];
