+++ src/types/topology/routes.rs
@@ -1,4 +1,5 @@
-use crate::types::topology::{NodeId, NodeRef};
+use crate::types::source_routing_header::NodeId;
+use crate::types::topology::topology::NodeRef;
 use std::collections::HashMap;
 
 pub struct Route {
